2019-12-26-Update-User-Logic

+++ src/main/java/com/michal/shopapp/service/UserService.java
@@ -3,9 +3,9 @@ package com.michal.shopapp.service;
 import com.michal.shopapp.components.dto.UserDTO;
 import com.michal.shopapp.components.mapper.UserMapper;
 import com.michal.shopapp.components.model.User;
+import com.michal.shopapp.exceptions.MyResourceNotFoundException;
 import com.michal.shopapp.repository.UserRepository;
-import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.michal.shopapp.utility.GlobalConstants;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -21,8 +21,10 @@ public class UserService {
         this.userMapper = userMapper;
     }
 
-    public Optional<UserDTO> findUserById(Long userId) {
-        return userRepository.findById(userId).map(userMapper::convertToDto);
+    public UserDTO findUserById(Long userId) {
+        return userRepository.findById(userId)
+                .map(userMapper::convertToDto).
+                        orElseThrow(() -> new MyResourceNotFoundException(GlobalConstants.USER_NOT_FOUND));
     }
 
     public UserDTO saveUser(UserDTO userDTO) {
@@ -30,7 +32,41 @@ public class UserService {
         return userMapper.convertToDto(savedUser);
     }
 
+    public UserDTO putUser(UserDTO userDTO, Long userID) {
+        Optional<User> userToUpdate = userRepository.findById(userID);
+        if (userToUpdate.isPresent()) {
+            userDTO.setId(userID);
+            User updatedUser = userRepository.save(userMapper.convertToEntity(userDTO));
+            return userMapper.convertToDto(updatedUser);
+        } else {
+            throw new MyResourceNotFoundException(GlobalConstants.USER_NOT_FOUND);
+        }
+    }
+
+    public UserDTO patchUser(UserDTO userDTO, Long userID) {
+        Optional<User> userToUpdate = userRepository.findById(userID);
+        if (userToUpdate.isPresent()) {
+            userToUpdate.ifPresent(u -> {
+                if (userDTO.getFirstName() == null) userDTO.setFirstName(u.getFirstName());
+                if (userDTO.getLastName() == null) userDTO.setLastName(u.getLastName());
+                if (userDTO.getBirthDate() == null) userDTO.setBirthDate(u.getBirthDate());
+                if (userDTO.getEmail() == null) userDTO.setEmail(u.getEmail());
+            });
+
+            userDTO.setId(userToUpdate.get().getId());
+            User savedUser = userRepository.save(userMapper.convertToEntity(userDTO));
+            return userMapper.convertToDto(savedUser);
+
+        } else {
+            throw new MyResourceNotFoundException(GlobalConstants.USER_NOT_FOUND);
+        }
+    }
+
     public void removeUserByID(Long userID) {
-        userRepository.deleteById(userID);
+        if (userRepository.existsById(userID)) {
+            userRepository.deleteById(userID);
+        } else {
+            throw new MyResourceNotFoundException(GlobalConstants.USER_NOT_FOUND);
+        }
     }
 }
